requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
dj-database-url==0.5.0
Django==4.0.6
django-on-heroku==1.1.2
psycopg2-binary==2.9.3
python-dotenv==0.20.0
sqlparse==0.4.2
whitenoise==6.2.0
